Borrado de POI e percorrido Asociar contas de Situm aos usuarios

+++ src/main/java/gal/caronte/sw/custom/PosicionCustom.java
@@ -5,14 +5,14 @@ public class PosicionCustom {
 	private Short idEdificio;
 	private Short idPlanta;
 	private Short nivel;
-	private Float latitude;
-	private Float lonxitude;
+	private Double latitude;
+	private Double lonxitude;
 	
 	public PosicionCustom() {
 		super();
 	}
 
-	public PosicionCustom(Short idEdificio, Short idPlanta, Short nivel, Float latitude, Float lonxitude) {
+	public PosicionCustom(Short idEdificio, Short idPlanta, Short nivel, Double latitude, Double lonxitude) {
 		super();
 		this.idEdificio = idEdificio;
 		this.idPlanta = idPlanta;
@@ -66,28 +66,28 @@ public class PosicionCustom {
 	/**
 	 * @return the latitude
 	 */
-	public Float getLatitude() {
+	public Double getLatitude() {
 		return this.latitude;
 	}
 
 	/**
 	 * @param latitude the latitude to set
 	 */
-	public void setLatitude(Float latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
 	/**
 	 * @return the lonxitude
 	 */
-	public Float getLonxitude() {
+	public Double getLonxitude() {
 		return this.lonxitude;
 	}
 
 	/**
 	 * @param lonxitude the lonxitude to set
 	 */
-	public void setLonxitude(Float lonxitude) {
+	public void setLonxitude(Double lonxitude) {
 		this.lonxitude = lonxitude;
 	}
