+++ src/main/java/gal/caronte/sw/modelo/puntointerese/PuntoInterese.java
@@ -20,11 +20,11 @@ public class PuntoInterese {
 	private Short idEdificio;
 	private Short idPlanta;
 	private Short nivel;
-	private Float latitude;
-	private Float lonxitude;
+	private Double latitude;
+	private Double lonxitude;
 
 	public PuntoInterese(Short idPuntoInterese, String nome, String descricion, Short idEdificio, Short idPlanta, Short nivel,
-			Float latitude, Float lonxitude) {
+			Double latitude, Double lonxitude) {
 		super();
 		this.idPuntoInterese = idPuntoInterese;
 		this.nome = nome;
@@ -123,28 +123,28 @@ public class PuntoInterese {
 	/**
 	 * @return the latitude
 	 */
-	public Float getLatitude() {
+	public Double getLatitude() {
 		return this.latitude;
 	}
 
 	/**
 	 * @param latitude the latitude to set
 	 */
-	public void setLatitude(Float latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
 	/**
 	 * @return the lonxitude
 	 */
-	public Float getLonxitude() {
+	public Double getLonxitude() {
 		return this.lonxitude;
 	}
 
 	/**
 	 * @param lonxitude the lonxitude to set
 	 */
-	public void setLonxitude(Float lonxitude) {
+	public void setLonxitude(Double lonxitude) {
 		this.lonxitude = lonxitude;
 	}
